Minor code organization
